[feature] extract auth form
