+++ glancy-site/src/components/index.js
@@ -1,2 +1,4 @@
 export { default as Button } from './Button/Button.jsx'
 export { default as ListItem } from './ListItem/ListItem.jsx'
+
+export { default as AuthForm } from "./AuthForm.jsx"
